Add login protection AND fix database error
I know, I did two things in one commit. Bad developer.

Added server logic that only allows logged-in users to see add, edit,
and delete pages for restaurants and menu items.

Also fixed server errors by creating unique sessions for each render
function in the server code. Was getting frequent "SQLite objects
created in a thread can only be used in that same thread" errors using
provided code with one master DB session for the entire server. Now I
don't get that error!

+++ project.py
@@ -8,11 +8,12 @@ from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem
 
-#Connect to Database and create database session
+# Connect to Database. I'll create unique sessions for each page, to avoid
+# database errors saying "SQLite objects created in a thread can only be used
+# in that same thread"
 engine = create_engine('sqlite:///restaurantmenu.db')
 Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
-session = DBSession()
 
 # Imports to implement user authentication
 # session already refers to a database session, so login sessions will be
@@ -172,19 +173,25 @@ def gdisconnect():
 #JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
 def restaurantMenuJSON(restaurant_id):
+    session = DBSession()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+    session.close()
     return jsonify(MenuItems=[i.serialize for i in items])
 
 
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
 def menuItemJSON(restaurant_id, menu_id):
+    session = DBSession()
     Menu_Item = session.query(MenuItem).filter_by(id = menu_id).one()
+    session.close()
     return jsonify(Menu_Item = Menu_Item.serialize)
 
 @app.route('/restaurant/JSON')
 def restaurantsJSON():
+    session = DBSession()
     restaurants = session.query(Restaurant).all()
+    session.close()
     return jsonify(restaurants= [r.serialize for r in restaurants])
 
 
@@ -192,73 +199,115 @@ def restaurantsJSON():
 @app.route('/')
 @app.route('/restaurant/')
 def showRestaurants():
-  restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+    session = DBSession()
+    restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
+    session.close()
+    return render_template('restaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
-  if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
-      session.add(newRestaurant)
-      flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-      session.commit()
-      return redirect(url_for('showRestaurants'))
-  else:
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to create a new restaurant')
+        return redirect(url_for('showRestaurants'))
+    if request.method == 'POST':
+        newRestaurant = Restaurant(name = request.form['name'])
+        session = DBSession()
+        session.add(newRestaurant)
+        flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+        session.commit()
+        session.close()
+        return redirect(url_for('showRestaurants'))
+    else:
       return render_template('newRestaurant.html')
 
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
-  editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to edit a restaurant')
+        return redirect(url_for('showRestaurants'))
+    # If logged in, go do stuff!
+    session = DBSession()
+    editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if request.method == 'POST':
       if request.form['name']:
         editedRestaurant.name = request.form['name']
+        session.add(editedRestaurant)
+        session.commit()
+        session.close()
         flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
         return redirect(url_for('showRestaurants'))
-  else:
-    return render_template('editRestaurant.html', restaurant = editedRestaurant)
+    else:
+        session.close()
+        return render_template('editRestaurant.html', restaurant = editedRestaurant)
 
 
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
-  restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-    session.delete(restaurantToDelete)
-    flash('%s Successfully Deleted' % restaurantToDelete.name)
-    session.commit()
-    return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
-  else:
-    return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to delete a restaurant')
+        return redirect(url_for('showRestaurants'))
+    # If logged in, go do stuff!
+    session = DBSession()
+    restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if request.method == 'POST':
+        session.delete(restaurantToDelete)
+        flash('%s Successfully Deleted' % restaurantToDelete.name)
+        session.commit()
+        session.close()
+        return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+    else:
+        session.close()
+        return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+
 
 #Show a restaurant menu
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
+    session = DBSession()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+    session.close()
     return render_template('menu.html', items = items, restaurant = restaurant)
-
 
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
-  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
-      session.add(newItem)
-      session.commit()
-      flash('New Menu %s Item Successfully Created' % (newItem.name))
-      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-  else:
-      return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to create a new menu item')
+        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+    # If logged in, go do stuff!
+    session = DBSession()
+    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if request.method == 'POST':
+        newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+        session.add(newItem)
+        session.commit()
+        session.close()
+        flash('New Menu %s Item Successfully Created' % (newItem.name))
+        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+    else:
+        session.close()
+        return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+
 
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to edit a menu item')
+        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+    # If logged in, go do stuff!
+    session = DBSession()
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     if request.method == 'POST':
@@ -272,26 +321,34 @@ def editMenuItem(restaurant_id, menu_id):
             editedItem.course = request.form['course']
         session.add(editedItem)
         session.commit()
+        session.close()
         flash('Menu Item Successfully Edited')
         return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
+        session.close()
         return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
+    if 'username' not in login_session:
+        # If nobody is logged in, you shouldn't be able to see this page!
+        flash('You must log in to delete a menu item')
+        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+    # If logged in, go do stuff!
+    session = DBSession()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
+        session.close()
         flash('Menu Item Successfully Deleted')
         return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
+        session.close()
         return render_template('deleteMenuItem.html', item = itemToDelete)
-
-
 
 
 if __name__ == '__main__':
